Move banner to home page only
