+++ apps/frontend/lib/frontend/web/templates/page/home/home.html.eex
@@ -1,6 +1,14 @@
-<iframe  class="kuikka-video" src="http://www.youtube.com/embed/xHRC67Isl28"
-  width="560" height="405" frameborder="0" allowfullscreen>
-</iframe>
+<div class="banner"
+  style="background-image: url('/images/kuikkaheader2.png'); display: block;">
+
+  <div class="next-operation">
+    <span class="u">Seuraava tapahtuma: <i class="tf"><%= get_next_event_date() %> 19:30</i></span>
+    <span class="b">Joka keskiviikko ja sunnuntai klo 19:30</span>
+  </div>
+  <iframe  class="kuikka-video" src="http://www.youtube.com/embed/xHRC67Isl28"
+    width="560" height="405" frameborder="0" allowfullscreen>
+  </iframe>
+</div>
 
 <div class="row align-middle">
   <div class="home-content-block">
@@ -30,7 +38,7 @@
     <h4>Ei arvonimiä</h4>
     <p>
 			Vaikka osa on kokeneempia pelaajia, ei heitä tarvitse sinutella emmekä vaadi
-			puhuttelu, esittely ja asia aloitessasi keskustelua...
+			puhuttelu, esittely ja asia aloittaessasi keskustelua...
     </p>
   </div>
 </div>
@@ -57,7 +65,7 @@
 <div class="row align-middle">
   <div class="home-content-block">
     <h2>Halukas liittymään kuikkaan?</h2>
-    <a href="/wiki/join">
+    <a href="">
       <button class="column home-join-button">Liity tästä</button>
     </a>
   </div>
