+++ apps/frontend/static/css/kuikka/_content.scss
@@ -1,19 +1,73 @@
-.fullwidth-banner {
+.banner {
   position: absolute;
-  top: 52px;
-  left: 0;
-  right: 0;
+  top: 50px;
+  left: -100%;
+  right: -100%;
   height: 505px;
   background-position: center;
   background-repeat: no-repeat;
-  background-size:cover;
+  background-size: cover;
+
+  .kuikka-video {
+    position: absolute;
+    top: 90px;
+    left: 40.5%;
+  }
+
+  .next-operation {
+    position: absolute;
+    left: 46%;
+    float: right;
+    margin-top: 15px;
+    background-color: rgba(215, 148, 0, 0.8);
+    padding: 0;
+    font-size: 28px;
+    font-weight: 100;
+    width: auto;
+    height: 40px;
+    opacity: 0.95;
+    color: #000;
+
+    span.u {
+      float: left;
+      display: block;
+      letter-spacing: 2px;
+      font-size: 28px;
+      margin: 0 0 2px 1px;
+      text-align: center;
+      line-height: 40px;
+      padding: 0 0 0 10px;
+
+      .tf {
+        color: #CCC;
+        font-style: normal;
+        background-color: #000;
+        padding: 0 10px;
+        float: right;
+        height: 40px;
+        line-height: 40px;
+        text-align: center;
+        margin-left: 10px;
+      }
+    }
+
+    span.b {
+      clear: both;
+      float: right;
+      width: auto;
+      display: block;
+      letter-spacing: 0;
+      font-size: 18px;
+      margin: -2px 0 0 0;
+      text-align: right;
+      line-height: 25px;
+      padding: 0 5px;
+      //background-color: #456116;
+      background-color: rgba(190, 118, 0, 0.8);
+    }
+  }
 }
 
-.kuikka-video {
-  position: absolute;
-  top: 150px;
-  left: 21%;
-}
 
 .home-content {
   margin-top: 145px;
@@ -52,56 +106,5 @@
   width: 100%;
 }
 
-.next-operation {
-	float: right;
-	margin-top: -480px;
-	//background-color: rgb(94, 128, 7);
-  background-color: rgba(215, 148, 0, 0.8);
-	padding: 0;
-	font-size: 28px;
-	font-weight: 100;
-	width: auto;
-	height: 40px;
-	opacity: 0.95;
-	color: #000;
-
-	span.u {
-		float: left;
-		display: block;
-		letter-spacing: 2px;
-		font-size: 28px;
-		margin: 0 0 2px 1px;
-		text-align: center;
-		line-height: 40px;
-		padding: 0 0 0 10px;
-
-		.tf {
-			color: #CCC;
-			font-style: normal;
-			background-color: #000;
-			padding: 0 10px;
-			float: right;
-			height: 40px;
-			line-height: 40px;
-			text-align: center;
-			margin-left: 10px;
-		}
-	}
-
-	span.b {
-		clear: both;
-		float: right;
-		width: auto;
-		display: block;
-		letter-spacing: 0;
-		font-size: 18px;
-		margin: -2px 0 0 0;
-		text-align: right;
-		line-height: 25px;
-		padding: 0 5px;
-		//background-color: #456116;
-  	background-color: rgba(190, 118, 0, 0.8);
-	}
-}
